Update dataflow.md
The explanation does not mention how to add the jar files into the "lib' folder. Furthermore I think that "libs" is a more appropriate naming for the folder containing all the libraries/jarfiles that is needed for the self-executing jar file to run.  See this page which explains how to add the libs dependencies inside the libs folder. The existing explanation assumes that we magically have a libs folder with al the dependencies. People with no experience creating jar files have problems understanding this and getting a beam job to run. Hope this change gets approved.